+++ CHANGELOG.md
@@ -5,4 +5,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - 2016-10-28
 ### Added
-- TODO
+- Development and production build processes using [Webpack](https://webpack.github.io/)